Recompile proto files.
The proto build script was broken with the introduction of Java chaincode.
PR #2162 resolved this and simplified the script.
Due to the script simplification, there were a few changes in the .pb.go
files after they were recompiled.
It was also necessary to change the package name of dah.proto
to avoid compile errors after compiling with the updated script.

Change-Id: Ie009c0b790db859e356645dd204082641ee7eab5
Signed-off-by: Sheehan Anderson <sheehan@us.ibm.com>