Run BDD Compose files intelligently
When bringing up containers with docker compose during BDD testing
a time.sleep call is not very efficient. This commit replaces this
with an intelligent probing of the containers to determine when they
are ready. It determines this by checking the output of the netstat
command on each container for listening or connected TCP connections.
Once an active connection is detected on all containers or the timeout
period elapses, execution continues or aborts respectively.

In combination with change 419 in this change BDDTests complete
approximately 4 minutes faster.

Signed-off-by: Julian Carrivick <cjulian@au1.ibm.com>
Change-Id: I2c89f360d04a8e5e5daeaa2aa5027e5c191a453c