Use repeated bytes instead of repeated string for chaincode call args
This allows applications to easily pass arbitrary blobs without having
to serialize them to strings.  At the same time, we also consolidate
the function argument to be part of the repeated bytes args.

For convenience and to simplify porting of existing chaincode to the
new argument format, we introduce helper functions in the shim which
cast between ([][]byte) and (string, []string).

Change-Id: I67562523a208727157c4767e86e1ef437e997f13
Signed-off-by: Gabor Hosszu <gabor@digitalasset.com>