Change APT to use mirror protocol
As part of the setup script, set APT to use the mirror protocol so
that no matter where you are in the world, the closest mirror to you
will be chosen. If you don't reside in the US it can be slow to
download all the packages when building images. This will mean the
best mirrors will be chosen regardless of the developer's location.

Change-Id: Iadda26f38aad8526219da2e4ef0fd66150004829
Signed-off-by: Julian Carrivick <cjulian@au1.ibm.com>