Move Docker-Compose files into their own folder
Move the BDD Docker compose files into a new sub folder so they are all
in one place and do not clog up the root of the bddtests/ folder.
While altering logic to determine the container names I also refactored
the Docker-Compose parsing code to make it more clear as to what it was
doing (parseComposeOutput in bdd_compose_util.py).

I needed to add a dummy compose-defaults.yml file so that CI would
successfully pass. When this change makes it into all branches then it
can be removed and the CI scripts updated.

I also changed the permissions on a couple of the bdd files from 755 to
644.

Signed-off-by: Julian Carrivick <cjulian@au1.ibm.com>
Change-Id: Iefd1007e79ee3827d39ed9b0bdec327f30e0c39f