Fix gmock links broken
This patch updated protobuf link from v3.0.0 to v3.0.2.
FAB-309 reported that there was no link to get gmock packages which was
required to build protobuf v3.0.0-beta-3. FAB-108 upgraded protobuf from
v3.0.0-beta-3 to v3.0.0. But the link to gmock packages was not fixed in
protobuf v3.0.0. This link was fixed in protobuf v3.0.2.

Fixes FAB-309

Change-Id: I20cf9674bd9e12ea3be5711ad719d18a6b8d9fd4
Signed-off-by: Nao Nishijima <nao.nishijima@hal.hitachi.com>