Add generic blackbox rawledger tests
This changeset adds tests which are generally applicable to
rawledger implementations.  It enables these tests for both
ramledger and for fileledger.

You may notice that these tests are substantially similar to
the tests in file ledger and were derived from there,
substituting black box equivalents for the implementation
details of fileledger.

This new suite of tests also caught a bug in the ramledger
implementation which could result in a crash when the
specified block was exactly the block number of the oldest
block in storage.  A fix is included.

Change-Id: I76813c8ba7853d47fa6ffa89b16ffc66b8ddfd5f
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>