Re-cut v0.6.0
This patch fixes two problems in commit e5a81bb6.

1) There was a problem where the #comment after the IS_RELEASE
variable was casuing the conditional logic to assume this was
a -snapshot release.

2) We collectively agreed that -developer-preview is a bit
of a mouthful for the release taxonomy and this should be
shortened to -preview

This has been confirmed to now drive the proper conditional
generation of a release tag and use the desired "-preview".

Change-Id: I58f2d68862aca34ae0181fe35c2f503e7f5d522a
Signed-off-by: Gregory Haskins <gregory.haskins@gmail.com>