Do not build the fabric peer during vagrant provision
Building the peer during vagrant provisioning takes
extra time. We can't be sure the user even wants to
build the peer.  They may be trying to perform some
other task (such as work on the SDK), in which case
we are wasting their time.  Therefore, remove peer
building from the critical path of provisioning
and allow the user to make their own decision.  This
has the nice benefit of making the provisioning
process nice and snappy.

Change-Id: I82f120faceeff8d8de08f064d4a6fd911eec65d9
Signed-off-by: Greg Haskins <gregory.haskins@gmail.com>