skeleton Endorser implemention with a CLI driver
This patch is for https://jira.hyperledger.org/browse/FAB-181,
FAB-182.

The patch is around endorser.go which implements the basic
Endorser service. The "peer deploy ..." and "peer invoke .."
CLI driver commands have been modified to redirected to use
endorser.ProcessProposal instead of the original "devops"
calls.

The deploy, invoke (and query) CLI calls are unchanged in
non-dev mode from user point of view - but for one difference.
The response is a ProposalResponse.

In dev mode (ie, when peer is started with --peer-chaincodedev)
the deploy command would need to set "CORE_CHAINCODE_MODE=dev"
in the "peer chaincode deploy ..." command. This is because, the
command now computes the chaincode code just like the SDK as
opposed to leaving it to be done by devops in the peer. Example
CORE_CHAINCODE_MODE=dev CORE_LOGGING_LEVEL=debug ./peer chaincode
deploy -n mycc -c '{"Args":["init","a","100","b","200"]}'

Change-Id: Ie6e44cef880bfcbeb7619f135566a7dce9dcdbc2
Signed-off-by: Srinivasan Muralidharan <muralisr@us.ibm.com>