Busywork updates 9/22/2016
Several changes are required due to the removal of the /devops route from
the REST API. The major change here is the conversion of the Tcl ::fabric
package to use the /chaincode route instead of /devops.

Several other enhancements and bug fixes are also inclued:

1) Fixes FAB-139 (Timer overflow)
2) Fixes FAB-414 (Additional timestamp for all TX committed)

Tcl support for reading /proc/<PID>/stat was added. This allows more
statistics to be printed after the Busywork run (e.g., user + system time
and CPU utilization). /proc/<PID>/stat fields can also be accessed for the
peers now by way of the 'busy' command, which was also enhanced with 'ps'
and 'pid' subcommands.

The 'pprofClient' was updated to support Go 1.7 naming conventions for pprof
files.

The -privacy flag was added to 'userModeNetwork'.

Removed trailing whitespace from all modified files.

Change-Id: Ie8d2cc95fb6e7384ce3fbb4d19949d14ec19ddc7
Signed-off-by: Bishop Brock <bcbrock@us.ibm.com>