Ported BDD implementations to golang using Godog.
Godog is the official Cucumber BDD framework for Golang,
which is the standard that our current behave (python) tests
are implemented.  In this manner, we reuse our existing feature files and can
implement BDD tests in both python and golang.

The Jira issue is in https://jira.hyperledger.org/browse/FAB-448

Change-Id: If95051670a0f4357f56b12216cd42f31b1402148
Signed-off-by: jeffgarratt <garratt.jeff@gmail.com>