Add string slice config support
When overriding a configuration option via the env which is supposed
to be a collection collection (slice) of options, the override is
instead interpretted as a string.  This patch modifies the config
decode hook to handle this special case and appropriately decode the env
override to a []string instead of a plain string.

There is no JIRA issue associated with this, but is being done to
support the Kafka epic effort, visible here:

https://jira.hyperledger.org/browse/FAB-32

Reclaim commit after a little Gerrit accident

Change-Id: Iae800b5237549232b90c7d7cf91590a3d8caf178
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>