Move orderer common components to common dir
As the complexity in the orderer grows, there is an increasing need to
add new component directories, which do not actually correspond to
different ordering services.

This can be seen with the bootstrap, cauthdsl, and broadcastfilter
directories, and more are on the way.

This changeset simply creates a new common directory, and moves the
common components there.  It also makes trivial updates to import paths,
but otherwise introduces no new features, so does not require new tests.

https://jira.hyperledger.org/browse/FAB-706

Change-Id: I61e803aa4454926b4ad26610680eb543554346b8
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>