Define Chain Configuration Protos
In order to bootstrap a chain, certain configuration must be embedded in
its genesis block.  This configuration needs to be in a well defined
format so that it is easy for consumers to read.  Further, this
configuration must be able to be modified in the future, so must define
policy which dictates how modification is allowed.

This changeset defines a set of protobuf messages in ab.proto which are
designed to satisfy the above requirements.  The configuration is
essentially a list of configuration items, which have a name (string), a
type (enum), a payload (bytes), and a reference to a modification policy
(string).  Additionally a sequence number scheme is included to prevent
replay.

This changeset also defines a wrapper for the policy configuration item
type.  It references the small proto DSL for signature validation, but
embeds this in a one-of concept in case anyone wishes to swap in other
policy validation schemes.

This is related to, but does not completely satisfy:

https://jira.hyperledger.org/browse/FAB-724

Change-Id: Id1cb73851a426f903455bb5b4eed552a65cc9327
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>