Hook config and policy pieces into orderer startup
This changeset takes the existing work of the config manager, policy
manager, and static genesis block creation and hooks them together into
the main execution path.

Note that the configuration manager is not utilized for any
configuration and there is no policy enforcement for chain
modifications, so this is only a half step, and will be followed by
utilization of the config manager and enforcement of policy on new
transactions.

Note also, that this changeset does not include tests by design, as it
is merely referencing existing pieces of tested code and adds no new
function itself.  All errors in this code path result in panic so the
bdd tests cover it as well.

Change-Id: Ie7f67971fee6c223378612af727e8f9aeca0e48e
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>