pull algorithm fix
Had a bug in the pull algorithm:
The NONCE set was actually needed to be split into 2 diff sets,
because servicing sync requests has nothing to do with
initiating them, and the cleanup of initiated sync requests
effected sync requets that were serviced.

Added a unit test that successfully reproduced the bug,
and after I fixed it it always succeeds.

Change-Id: Ie99f41485a41ddb46e99d8fa7d8db408f669b20c
Signed-off-by: Yacov Manevich <yacovm@il.ibm.com>