Orderer Top Level Message Format
This is the a proposed message format for the ordering service. This
format will be used for all broadcast messages, and all blocks will
consist of marshaled orderer messages of type Envelope.

The fabric will extend these messages by depending on the header for
some information that was previously repeated inside the transaction,
and having a second inner header per proposal.

Change-Id: Id764490f5443e893ea5f8f98c60367a53715e272
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>