BCCSP Factory support
This change-set introduces support for multiple BCCSP factories.
This way, it will be possible to have software-based as well as
hardware-based BCCSP implementations.
The fatory provides also access to a default BCCSP that represents the BCCSP
to be used by default as it is configured via viper.
Tests have been added to test the factories.
This change-set comes in the context of:
https://jira.hyperledger.org/browse/FAB-354

Change-Id: Ib83f1fc31fa0ac77daa76f97684563e4c6e53877
Signed-off-by: Angelo De Caro <adc@zurich.ibm.com>