FAB-959 Chaincode with binary data in CouchDB
CouchDB is most suited for handling chaincode with JSON values.
But in some cases chaincode with binary values will need to be stored
in CouchDB state database, for example to support chaincode
deployment in lifecycle chaincode.

This change adds attachment support for couchDB in order to handle
chaincode with binary data, in addition to chaincode with JSON data.

For this change, only one attachment is supported even though
the attachment is added as an array.

The new signature for SaveDoc now allows for either a JSON represented
as a byte array or an array of attachments.

Change-Id: Ifac9ce56c053ed929dae2aab4a5932c0453680a7
Signed-off-by: Chris Elder <chris.elder@us.ibm.com>