[BUILD] Refactor make to simplify rules
It is not strictly necessary to create the -src base image
since we can just use a volume mount in its place. There is
also quite a bit of repetition in the build process right
now.  Therefore, this patch refactors the build to eliminate
the -src image as well as consolidates/simplifies a few things.

This sets the stage for more precise Dockerfile tunings
(such as EXPOSE/CMD on a per-daemon basis)

Change-Id: Ib0ff8e461eb4ba661b201eb519609b3f35ceab13
Signed-off-by: Greg Haskins <gregory.haskins@gmail.com>