[FAB-798] Abstract out the solo deliver handler
As the next step of consolidating the common logic of the
atomicbroadcast api between components, this changeset pulls out the
logic which is not solo specific and moves it into the common/deliver
package.

SBFT was already dependent on this functionality, so this is a natural
move.

This continues, but does not satisfy FAB-798.

Change-Id: I02c5ef5b03f9e1a17fd188e3df7b6fb42aa126b5
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>