[FAB-798] Factor common gRPC components from solo
As the latest step in consolidating the common logic between orderers,
this changeset removes the gRPC components from solo, and distills solo
to its most simple core of a 'solo consenter'.

The solo consenter performs batch cutting and ordering, while the
broadcast filtering and deliver logic was removed in previous
changesets.

Change-Id: I5f9dfe6239356279a9625a2c89d327c5da275bbc
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>