Integration of MSP in endorser
This code change integrates the MSP into the peer, endorser and ESCC. Currently
the MSP is used to validate the proposal (in the endorser) and endorse it (in
the ESCC). Also, the peer CLI uses the MSP to sign the proposal to be sent
to the endorser. Two MSP implementations are available: a no-op and one that
uses BCCSP (and golang native x509 support) to handle x509 certificates.

Missing:
 * loading the roots of trust from the genesis block using a CSCC call;
 * loading more than one identity;
 * specifying the location of the config file in the .yaml config;
 * a way of identifying signing identities for the CLI;

NOTE: this commit also contains some code from
 - https://gerrit.hyperledger.org/r/#/c/2243/
 - https://gerrit.hyperledger.org/r/#/c/2385/
 - https://gerrit.hyperledger.org/r/#/c/2465/

Change-Id: I831b7bd01eb8b71cd3a6d9893a4faf6f7fa90fed
Signed-off-by: Alessandro Sorniotti <ale.linux@sopit.net>
Signed-off-by: Elli Androulaki <lli@zurich.ibm.com>