[FAB-996] Introduce orderer/commons/util package
https://jira.hyperledger.org/browse/FAB-996

The goal is to avoid rewriting the same functions again and again.

For functions that may return an error, two variants and provided:
- The plain one that returns the error and allows the caller to handle it
- The *OrPanic one that panics if an error comes up. It makes sense to use
  the *OrPanic functions in contexts where you either know the chances of
  an error are zero, or you're willing to defer/recover in the calling
  code. The static package is a perfect example of this. (Also see note
  below.) The *OrPanic functions can be used inline as function arguments
  since they return a single object.

This changeset ripples into the static bootstrapper, simplifying it
substantially, and providing a clear guide into creating a sample genesis
block.

The Make* functions of the util package are tested in the static
bootstrapper's tests. Unit tests for functions that just create proto
messages don't make sense, as any errors in the message construction
results in immediate 'go build' errors. (All the other functions in the
util package come with unit tests, as they should.)

UPDATE: Revised for the latest changes in protos.

Change-Id: I4212d1366cca556fe4c5c6cb16ea3c1f52dd7503
Signed-off-by: Kostas Christidis <kostas@christidis.io>