sbft: rework new view null requests
Previously we would use our own concept of LastBatch() for the null
request batch prevHash.  This was unsound and replicas that were in view
change for one or more requests could not verify the correctness of this
hash.  Instead we now base our choice of prevHash on the now properly
reported most recent checkpoint.

Change-Id: Ia6794113672d64700b05ed10cac8e81539428f5b
Signed-off-by: Simon Schubert <sis@zurich.ibm.com>