Change how chaintool executes
The chaintool utiliy is made executable by pre-pending
some bash statements within the chaintool binary itself.
One of the parameters used to make this happen is
-Xbootclasspath/a.

Currently, the IBM JVM on the s390x platform cannot
process the -Xbootclasspath/a parameter, and as a
result, the following error message is diplayed:
"Could not find or load main class chaintool.core".

To allow all platforms to work with the package.go
file, my patch invokes the chaintool using the command:
"java -jar /usr/local/bin/chaintool"

Change-Id: I164b460cc2d2dbcaca7a102a8aa31de8fdc2751d
Signed-off-by: John Harrison <jharriso@us.ibm.com>