TX proposal/endorsement/validation flow (+MSP)
This change set contains a set of functions to generate a transaction (from
proposal, endorsements and a signing identity) and validate it (given a set
of root CAs). The validation code will be used by the committer. The tx
assembling code should be helpful for the SDK team to understand how
transactions should be assembled. Additionally, it has changed the type of
messages exchanged everywhere to be of the proper type and with signatures
(obtained from a fixed identity for now). Finally, it contains an initial
implementation of VSCC with unit tests (which is however not yet called by
the committer).

Change-Id: I375ecc7e61516f3c4ab8fd874aa564e99cc720fb
Signed-off-by: Alessandro Sorniotti <ale.linux@sopit.net>