Remove non-determinism in tests (undo FAB-839)
The changeset submitted for FAB-839 introduces some non-determinism which
makes the tests in the kafka package block. To test that this is the
case, checkout that changeset [1] and run the tests in the kafka package
with the "-count 50" option. A quick inspection shows we may be dealing
with context leakage, but I'll defer to Luis to investigate this further.
Until then, this changeset disables the new test and restores the old
send-replies logic of the code. It passes all tests.

[1] https://gerrit.hyperledger.org/r/#/c/2043/

Change-Id: Ie29bc38288fad3c893bc332771974ec1f19b7a4b
Signed-off-by: Kostas Christidis <kostas@christidis.io>