[DEVENV] Install docker-compose v1.8.1 locally
We need docker-compose v1.8.1 but baseimage v0.2.1 includes
v1.5.2.  We will eventually be migrating away from the
vagrant baseimage anyway, so we are going to pretend
docker-compose is not installed regardless of whether
it might be or not.  We will instead plow it over with
the release that we require.

Users will need to vagrant destroy+up after this patch

Change-Id: I7ae5eb87953466fadf1a7b5fb272633c35f67898
Signed-off-by: Gregory Haskins <gregory.haskins@gmail.com>