[FAB-421] Add multi-chain support to rawledger
The original rawledger interface only supports a single chain.  However,
to support multiple chains, this interface must be extended to support
multiple chains.  This changeset creates a Factory interface for the
rawledger which supports the method GetOrCreate which will get the
ledger if it already exists, or create a new one if not.

In order to preserve some backwards compatability, and to support the
notion of multi-chain more generally in the future with an ordering
system chain, the factories accept a genesis block to seed the ordering
system chain with.

Change-Id: I99844ff1b9a11f383d08c23f60d95f848632876e
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>