Create multichain manager
The multichain manager will ultimately supplant much of what is main.go
and generally coordinate the creation of the consenter backed chains and
provide an interface for the common broadcast/deliver logic.

This changeset introduces the manager but makes no attempt to hook it
into the rest of the system (for ease of review).

Change-Id: I89072ff497dbf0395f1864587e034d5906217cea
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>
Signed-off-by: Yacov Manevich <yacovm@il.ibm.com>