BCCSP KeyStore
This change-set introduce the concept of KeyStore at the BCCSP level.
A KeyStore represents a storage system for cryptographic keys.
It allows to store and retrieve bccsp.Key objects. The key store can be
read only, in that case storing a key will return an error.
Now, to initialize a software-based BCCSP, a key store must be specified.
A file based keystore is provided.

In addition, all the dependencies from viper has been removed.
In order to do that, some assumptions has been made on default
security level. The organization of this paramenters will be better
addressed in a subsequent change-set.

This change-set comes in the context of:
https://jira.hyperledger.org/browse/FAB-354

Change-Id: I7e96eea7e5e89ecec86863ebc11b919d29c0019a
Signed-off-by: Angelo De Caro <adc@zurich.ibm.com>