Rename orderer/config to orderer/localconfig
The orderer service has two different types of configuration,
configuration which is local to the process (such as buffer sizes, and
backing structures like ledger type and location) as well as global
configuration which must be the same across nodes, such as the consensus
type and batch size.

Today, this information is all stored in one local config file
(orderer.yaml) but this needs to change.  As a first step to
differentiate between config which is local and which is shared, this
changeset renames the current yaml config mechanism to be localconfig
with the expectation that a sharedconfig package will be created in the
future.

Change-Id: I72298554600d8dc0d0c441e47fe9ab9c801e6aab
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>