point COMMITTER_LEDGER_ORDERER to right default
The right default is "0.0.0.0:7050". This is
needed for usability so users of the latest
master code have fewer things to keep track of
when running with defaults.

Note that in core.yaml REST also points to 7050.
This is not an issue as REST is is no longer used.
Will leave that in for now pending a larger
core.yaml cleanup.

Change-Id: Iac64fcb8b680a7b8a47b5c105da628c999ad06f0
Signed-off-by: Srinivasan Muralidharan <muralisr@us.ibm.com>