Introduce orderer sharedconfig mechanism
The configtx.Manager supports maintaining configuration updates based on
configuration transactions, but today, the only interpretation of these
configuration updates is the policies.Manager which handles
configuration updates of type Policy.

There are other types of configuration, such as Orderer type
configuration which are currently ignored.

This changeset adds an orderer shared config manager.  It is called
sharedconfig to make it explicit that this configuration is common to
and available to all orderers.  This is different from the notion of
localconfig which could include things like machine dependent paths and
buffer sizes.

This changeset also establishes a convention for encoding orderer
configuration by creating a proto message with name equal to the Key
field which is set in the orderer config.

This changeset hooks the shared config into the XXX points of the
multichain ChainSupport.  In the future, the batchSize and
consensusType will be removed from the localconfig once the other
consensus plugins have been migrated to the new shared multichain
infrastructure.

Change-Id: Ia68f5a83bf816317609b7df4c9ad0d1bed1bb800
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>