[FAB-816] Clean multichain integration interfaces
The multichain integration in https://gerrit.hyperledger.org/r/#/c/2721/
hooked the multi-chain manager into the default solo path.  This was
attempted in a way which kept the diff minimal, but correspondingly
produced some unpleasant artifacts in the code.

In particular, the integration used one catch-all interface in
multichain.Manager to supply support to the various components, this
meant that components were forced to drag in unnecessary imports and it
complicated mock testing with many panic-ing unimplemented functions.

This changeset breaks this interface into pieces, and pushes the
definition of the interface back into the components which depend on the
definition.

Also included in this changeset is substantial cleanup of the
solo/consenter_test.go file. This set of tests was originally written in
a way which depended on the end to end flow of the system, but as the
common components have been factored out, and as the solo 'consensus'
has been reduced to its simplest component, more targetted tests are now
needed.

Also included are some assorted linting fixes.

Change-Id: I7c5e20cd7b8c66eb51cc56ad539177ce81cbcbfc
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>