Re-enable configtx handling
Before the orderer and peer components standardized on a proto message
format for shared communication, some proto messages were being
unmarshaled as the wrong type, resulting in messages which looked like
configuration transactions, but were not.  All these transactions were
then rejected because they did not apply correctly as configuration.

The fabric has since standardized on a shared set of protos, so it
should be safe to re-enable configuration transaction filtering.

This changeset re-enables the configuration filters.

Additionally, this changeset hooks the configuration transaction
execution (application) into the multichain manager writerinterceptor
path and adds tests for this path.

Change-Id: I5ba75861812d54e19fa6eeac98174215761f9ad2
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>