Suppress logging output of the peer during unit-tests
We recently introduced the usage of docker-compose to
orchestrate the unit-test execution.  In doing so, we
changed the behavior of the peer container to inadvertently
make the peer log visible.  This isn't strictly necessary,
and it pollutes the unit-test output.  Therefore, lets
configure docker-compose to act similar to the previous
setup by suppressing the peer.

Change-Id: I2e331d6ab97e330ada89de9af0f690406324506d
Signed-off-by: Gregory Haskins <gregory.haskins@gmail.com>