FAB-1200 wrong type assertion on ledger.KV
https://jira.hyperledger.org/browse/FAB-1200

While investing table structures (still in works) found that
chaincode handler was asserting the wrong KV type.

Also found that an UnmarshallJSON was missing on proto.ChaincodeInput.
This function was defined in transaction.go (wrong file for a
ChaincodeInput function!) which went away when old Transaction was
axed. Interesting that the tests in common_test.go did not catch that.
Added explicit tests for Unmarshalling JSON into ChaincodeInput.

Change-Id: I3adc92c888efee974c5f537bc911cb0a253d1e01
Signed-off-by: Srinivasan Muralidharan <muralisr@us.ibm.com>