[FAB-1278] Generalize Config Execution Path
For the impending chain creation work, a second type of transaction will
need to be 'executed' beyond a configuration transaction.  This
changeset generalizes the old configuration specific code paths into a
re-usable path.

In particular, the broadcast filters have been generalized to be a more
generic filtering mechanism.  Instead of replying with the matched rule
type, and then having the invoker make decisions based on the match, the
filters now return a Committer which can perform those actions with no
specific knowledge from the caller.

The configuration transaction was being treated specially in several
parts of the orderer code.  This created ugly case switching and
ultimately produced more code than with a more generic approach.  So
this changeset produces a net decrease in overall lines of code.

Change-Id: Ifaa769a0303f828de1d76e0a014e47d9044756ce
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>