[FAB-814] Introduce ChainCreators orderer config
This changeset adds an orderer configuration item with key
ChainCreators.  This configuration item should only be set for the
system chain, and will allow the specification of chain creation
policies for an orderer.

This changeset simply introduces the new configuration item but does not
utilize it in any way (so as to simplify the diff).  An additional
changeset follows which utilizes this new config.

This is the server side half of the FAB-814 item but does not complete
it.

Change-Id: Ia761d4a05d5a89875a99b30210744f31ad7112d5
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>