FAB-829: App library for access control/App. MSP
This is a changeset including some building blocks for the
application (chaincode invoker/chaincode) to prove ownership
of certain attributes and the chaincode to certify such
ownership proof.

The changeset includes also an interface for a server-aided
MSP that includes the online registration and  enroll
process of an entity with the MSP server.

Building blocks are based on peermsp for now.

Change-Id: Id42a07a14680f7961fd9bb87e54db913ebff30b8
Signed-off-by: Elli Androulaki <lli@zurich.ibm.com>