Invoke VSCC from committer
This change-set adds the call to VSCC to the committer's code path. There are
still a few missing pieces that we can only address once LCCC is finalized
(e.g. identify the right VSCC, extract policies). The change-set also contains
a minor refactoring of the tx-assembly code as discussed with Manish.

Change-Id: Ibf8807dbb2934dcf4929c3e1a3edd4e5ec9f40ea
Signed-off-by: Alessandro Sorniotti <ale.linux@sopit.net>