[FAB-1280] Create fabric common components
https://jira.hyperledger.org/browse/FAB-1280

There are components in the orderer/common directory which are of more
general use outside of the orderer, such as the cauthdsl, configtx, and
policies components.

Because there are likely other useful common components, such as the
impending gRPC server, logging, and msp/bccsp work, this changeset
creates a fabric/common directory and moves the above listed components
from the orderer to this directory.

Change-Id: I01765f1833685f768079211542f3b0ab4b00ed0e
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>