Switch blockcutter to always use sharedconfig
There was a XXX in the multichain manager to push the blockcutter
batchSize into the blockcutter via sharedconfig, rather than to pass it
at construction time.  This is necessary to be able to dynamically
change the value of the batch size after genesis.

This changeset also introduces a sharedconfig mock structure in the
orderer/mocks package as it will likely be more broadly useful to other
tests going forward.

Change-Id: I1af47c3ca62a545236efa1e03280203ec5e75a76
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>